requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic-settings
pymupdf

langchain
langchain-core
langchain-community
langchain-google-genai>=2.0.0
langchain-text-splitters

chromadb
google-genai
python-dotenv